ADD:  coreData 엑세스 담당하는 LocalStorageServiceImpl 클래스 생성
